backend/requirements.txt
asgiref==3.11.0
celery==5.4.0
Django==6.0.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
PyJWT==2.10.1
python-dotenv==1.2.1
redis==5.2.1
requests==2.32.3
sqlparse==0.5.5
mailtrap
user-agents==2.2.0
